python/cookbooks/code_gen_agent_ax/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
arize==7.51.2
arize-otel==0.11.0
attrs==25.4.0
beautifulsoup4==4.12.3
brotli==1.2.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.124.2
ffmpy==1.0.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.12.0
googleapis-common-protos==1.72.0
gradio==6.1.0
gradio-client==2.0.1
groovy==0.1.2
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==1.2.2
idna==3.11
importlib-metadata==8.7.0
jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.66
langchain-openai==0.3.19
langchain-text-splitters==0.3.8
langgraph==0.2.62
langgraph-checkpoint==2.1.2
langgraph-sdk==0.1.74
langsmith==0.4.58
loguru==0.7.2
markdown-it-py==4.0.0
markupsafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.7.0
mypy-extensions==1.1.0
numpy==2.3.5
openai==1.109.1
openinference-instrumentation==0.1.42
openinference-instrumentation-langchain==0.1.56
openinference-instrumentation-openai==0.1.41
openinference-semantic-conventions==0.1.25
opentelemetry-api==1.39.0
opentelemetry-exporter-otlp==1.39.0
opentelemetry-exporter-otlp-proto-common==1.39.0
opentelemetry-exporter-otlp-proto-grpc==1.39.0
opentelemetry-exporter-otlp-proto-http==1.39.0
opentelemetry-instrumentation==0.60b0
opentelemetry-proto==1.39.0
opentelemetry-sdk==1.39.0
opentelemetry-semantic-conventions==0.60b0
orjson==3.11.5
ormsgpack==1.12.0
packaging==24.2
pandas==2.2.3
pillow==12.0.0
propcache==0.4.1
protobuf==6.33.2
pyarrow==22.0.0
pydantic==2.12.4
pydantic-core==2.41.5
pydantic-settings==2.12.0
pydub==0.25.1
pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.3
regex==2025.11.3
requests==2.32.5
requests-futures==1.0.0
requests-toolbelt==1.0.0
rich==14.2.0
safehttpx==0.1.7
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlalchemy==2.0.45
starlette==0.50.0
tenacity==9.1.2
tiktoken==0.12.0
tomlkit==0.13.3
tqdm==4.66.6
typer==0.20.0
typer-slim==0.20.0
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.6.1
uuid-utils==0.12.0
uvicorn==0.38.0
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0